% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/potentials.R
\name{as.matrix.potential}
\alias{as.matrix.potential}
\title{as.matrix.potential}
\usage{
\method{as.matrix}{potential}(obj)
}
\arguments{
\item{obj}{potential obj}
}
\description{
Get potential matrix
}

