% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentData.R
\docType{data}
\name{BammTransProb}
\alias{BammTransProb}
\title{BaMM transition probabilities for CTCF Motif}
\format{A list of transition matrices.}
\source{
\url{https://github.com/soedinglab/BaMMmotif} \url{http://hgdownload.cse.ucsc.edu/goldenPath/hg19/encodeDCC/wgEncodeAwgTfbsUniform/}
}
\usage{
BammTransProb
}
\description{
A 2nd order Markov chain representation of the 
CTCF motif. The parameters are obtained using the
BaMM software and using ENCODE ChIP-seq peaks for the
CTCF binding site in MCF-7 cells.
}
\details{
See also the vignette BaMM Analysis.
}
\keyword{datasets}

