% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/dfg.R
\name{dfg}
\alias{dfg}
\title{dfg
Construct a probalistic graphical model}
\usage{
dfg(varDim, facPot, facNbs, varNames = seq_along(varDim),
  facNames = (length(varDim) + seq_along(facPot)))
}
\arguments{
\item{varDim}{Vector of integers containing variable dimensions}

\item{facPot}{List of matrices containing factor potentials for null model}

\item{facNbs}{List of vector containing describing neighbouring variables for each factor}
}
\description{
dfg
Construct a probalistic graphical model
}
\examples{
varDim <- c(2L,3L)
facPot <- list(matrix(c(1:6),2,3))
facNbs <- list( c(0L,1L))
my_pgm <- dfg(varDim, facPot, facNbs)
}

