\name{dgRaph-package}
\alias{dgRaph-package}
\alias{dgRaph}
\docType{package}
\title{
Inference in Factor Graphs
}
\description{
Training and evaluation of factor graphs.
}
\details{
\tabular{ll}{
Package: \tab dgRaph\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-06\cr
License: \tab GPL-3 \cr
}
}
\author{
Tobias Madsen

Maintainer: Tobias Madsen <tobias.madsen@gmail.com>
}
\references{
For an introduction to factor graphs see chapter 8 in \emph{Pattern Recognition and Machine Learning, Bishop. C., 2007 }
}
\keyword{ package }
\seealso{
\code{\link[dgRaph:dfg]{dfg}}
}
