% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect.R
\name{expect.dfg}
\alias{expect.dfg}
\title{Expectations in DFG}
\usage{
\method{expect}{dfg}(x, facScores, data = NULL, ...)
}
\arguments{
\item{x}{dfg object}

\item{facScores}{A list of matrices similar to factor potentials indicating a function which expectation should be evaluated}

\item{data}{dataframe or matrix with observed data. The mapping between columns and rows will be performed automatically. NA will be interpreted as a missing variable}

\item{...}{Arguments to be passed to methods.}
}
\description{
Expectations in DFG
}

