% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expect.R
\name{expect}
\alias{expect}
\title{Expectation
Generic function for finding expectancy of function under a distribution x}
\usage{
expect(x, ...)
}
\arguments{
\item{x}{distribution under which the expecation should be evaluated}

\item{...}{Arguments to be passed to methods.}
}
\description{
Expectation
Generic function for finding expectancy of function under a distribution x
}

