% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect.R
\name{expect2.dfg}
\alias{expect2.dfg}
\title{Expectations and 2nd order moments in DFG}
\usage{
\method{expect2}{dfg}(x, facScores, ...)
}
\arguments{
\item{x}{dfg object}

\item{facScores}{A list of matrices similar to factor potentials indicating a function which expectation should be evaluated}

\item{...}{Arguments to be passed to methods.}
}
\description{
Expectations and 2nd order moments in DFG
}

