% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect.R
\name{expect2}
\alias{expect2}
\title{Expectation 2nd order moment
Generic function for finding likelihood, expectancy and 2nd order moment of function under a distribution x}
\usage{
expect2(x, ...)
}
\arguments{
\item{x}{distribution under which the expecation should be evaluated}

\item{...}{Arguments to be passed to methods.}
}
\description{
Expectation 2nd order moment
Generic function for finding likelihood, expectancy and 2nd order moment of function under a distribution x
}

