% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optimize.R
\name{fixedlinkOptimize}
\alias{fixedlinkOptimize}
\title{Fixed link optimization
Known linear relationship between two variables but unknown variance}
\usage{
fixedlinkOptimize(range1 = c(0, 100), range2 = c(0, 100), alpha = 1,
  beta = 0)
}
\arguments{
\item{range1}{range of independent variable}

\item{range2}{range of dependent variable}

\item{alpha}{known slope}

\item{beta}{known intercept}
}
\description{
Fixed link optimization
Known linear relationship between two variables but unknown variance
}

