% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/potentials.R
\name{fixedNormalPotential}
\alias{fixedNormalPotential}
\title{Fixed Normal Potential}
\usage{
fixedNormalPotential(dim = c(100, 100), range1 = c(0, 100), range2 = c(0,
  100), a = 1, b = 0, sd = diff(range2)/5)
}
\arguments{
\item{dim}{A vector with dimensions of potential}

\item{range1}{The limits of binning scheme 1st variable}

\item{range2}{The limits of binning scheme 2nd variable}

\item{a}{Slope}

\item{b}{Intercept}

\item{sd}{standard deviation}
}
\description{
Fixed Normal Potential
}

