% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.properties.R
\name{getwtkcore}
\alias{getwtkcore}
\title{Weighted k-core induced subgraph}
\usage{
getwtkcore(graph, mode = c("all", "out", "in"), k, weights = NULL,
  type = c("first", "second"), alpha = NULL, beta = NULL, lambda = NULL,
  wtdeg = c("once", "rep"))
}
\arguments{
\item{graph}{The input graph, it can be directed or undirected}

\item{mode}{The type of the core in directed graphs. Character constant,
possible values: \code{in}: in-cores are computed, \code{out}: out-cores are
computed, \code{all}: the corresponding undirected graph is considered. This
argument is ignored for undirected graphs.}

\item{k}{The core number for which the subgraph has to be induced.}
}
\value{
Graph object of the induced subgraph
}
\description{
The k-core of graph is a maximal subgraph in which each vertex has at least
degree k.
}
\details{
The k-core induced subgraph is the maximal subgraph in which every vertex has at
least degree k.

This function calculates the induced subgraph for core 'k'.
}
\examples{

g <- make_ring(10)
g <- add_edges(g, c(1,2, 2,3, 1,3))
getkcore(g,k=3) 		# small core triangle in a ring

}
\author{
Chhaya Malik \email{chhaya.mcs.du.2015@gmail.com}

Mani Bansal \email{mani.mcs.du.2015@gmail.com}
}
\references{
Vladimir Batagelj, Matjaz Zaversnik: An O(m) Algorithm for Cores
Decomposition of Networks, 2002

Seidman S. B. (1983) Network structure and minimum degree, \emph{Social
Networks}, 5, 269--287.
}
\seealso{
\code{\link{coreness}}
}
\keyword{graphs}

