% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/kl.R
\name{kl}
\alias{kl}
\title{Kullback-Leibler divergence}
\usage{
kl(dfg1, dfg2)
}
\arguments{
\item{dfg1}{Discrete factor graph object describing background distribution}

\item{dfg2}{Discrete factor graph object describing foreground distribution}
}
\description{
Calculates Kullback-leibler divergence KL(dfg1, dfg2)
between two factor grahs with identical structure.
}

