% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\title{Calculate likelihood of observation}
\usage{
likelihood(data, dfg, log = FALSE, dataList = list())
}
\arguments{
\item{data}{dataframe or matrix with observed data. The mapping between columns and rows will be performed automatically. NA will be interpreted as a missing variable}

\item{dfg}{discrete factor graph object}

\item{log}{calculate loglikelihood}
}
\value{
A vector of likelihoods for each observation
}
\description{
Calculate likelihood of observation
}

