% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/potentials.R
\name{linregPotential}
\alias{linregPotential}
\title{Linear Regression Potential
Initialize a linear regression to reasonable defaults}
\usage{
linregPotential(dim = c(100, 100))
}
\arguments{
\item{dim}{A vector with dimensions of potential}
}
\description{
Linear Regression Potential
Initialize a linear regression to reasonable defaults
}

