% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/potentials.R
\name{logregPotential}
\alias{logregPotential}
\title{Log Regression Potential}
\usage{
logregPotential(dim = c(100, 100), range1 = c(0, 100), range2 = c(0, 100),
  alpha = NULL, beta = NULL, var = NULL)
}
\arguments{
\item{dim}{A vector with dimensions of potential}

\item{range1}{Range of independent variable}

\item{range2}{Range of dependent variable}

\item{alpha}{Slope of linear regression}

\item{beta}{Intercept of linear regression}

\item{var}{Variance}
}
\description{
Initialize a log regression to reasonable (random) defaults.
\deqn{y = \alpha \log(x) + \beta + \epsilon}
If alpha, beta and var is provided they will be used
}

