% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanlinkPotential.R
\name{meanlinkPotential}
\alias{meanlinkPotential}
\title{Mean link potential}
\usage{
meanlinkPotential(dim = c(100, 100), range1 = c(0, 100), range2 = c(0,
  100), alpha = 1, beta = 0, var = NULL)
}
\arguments{
\item{dim}{A vector with dimensions of potential}

\item{range1}{Range of independent variable}

\item{range2}{Range of dependent variable}

\item{alpha}{Slope of linear regression}

\item{beta}{Intercept of linear regression}

\item{var}{Variance}
}
\description{
Initialize a meanlink potential. The first neighbour encodes the conditional mean in the second neighbour. The variance can still be learned from data. Internally a linregPotential where alpha and beta are fixed trough when training.
}

