% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/mps.R
\name{mps}
\alias{mps}
\title{Most Probable State}
\usage{
mps(data, dfg)
}
\arguments{
\item{data}{dataframe or matrix with observed data. The mapping between columns and rows will be performed automatically. NA will be interpreted as a missing variable}

\item{x}{points to evaluate tail probabilities in}
}
\value{
A dataframe with a column for each variable and the most probable state for each observation
}
\description{
Most Probable State
}

