% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/potentials.R
\name{multinomialPotential}
\alias{multinomialPotential}
\title{Multinomial Potential
Initialize a multinomial potential with random starting point}
\usage{
multinomialPotential(dim = c(1, 5))
}
\arguments{
\item{dim}{A vector with dimensions of potential}
}
\description{
Multinomial Potential
Initialize a multinomial potential with random starting point
}

