% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/potentials.R
\name{normalPotential}
\alias{normalPotential}
\title{Normal Potential
Initialize a norm potential with random starting points}
\usage{
normalPotential(dim = c(1, 100))
}
\arguments{
\item{dim}{A vector with dimensions of potential}
}
\description{
Normal Potential
Initialize a norm potential with random starting points
}

