% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentData.R
\docType{data}
\name{PFM}
\alias{PFM}
\title{Position Frequency Matrix for CTCF Motif}
\format{A 4x19 matrix giving the nucleotide frequency 
at each of 19 positions.}
\source{
\url{http://jaspar.genereg.net/}
}
\usage{
PFM
}
\description{
A Position Frequency Matrix (PFM) for the CTCF Motif
obatined from the Jaspar motif database.
}
\keyword{datasets}

