% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dfg.R
\name{plot.dfg}
\alias{plot.dfg}
\title{plot.dfg}
\usage{
\method{plot}{dfg}(x, ...)
}
\arguments{
\item{x}{dfg object to plot}

\item{...}{Arguments to be passed to plotting method. Typically \code{layout}, for connected graphs \code{layout = layout.reingold.tilford} gives good results.}
}
\description{
Plot the graphical structure of a dfg object
}
\examples{
library(dgRaph)
varDim <- rep(4,5)
facPot <- list(multinomialPotential(c(1,4)),
               multinomialPotential(c(4,4)))
facNbs <- list(c(1L),
               c(1L,2L),
               c(1L,3L),
               c(3L,4L),
               c(3L,5L))
potMap <- c(1,2,2,2,2)
facNames <- c("P",rep("I",4))
varNames <- c("Do","Re","Mi","Fa","Sol")
mydfg <- dfg(varDim, facPot, facNbs, potMap, varNames, facNames)
plot(mydfg)
plot(mydfg, layout = layout.reingold.tilford)
}

