% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poibin.R
\name{poibinSA}
\alias{poibinSA}
\title{Poison binomial tail probabilities using Saddlepoint Approximation}
\usage{
poibinSA(t, p, s = rep(1, length(p)), lattice = 1L, log = T)
}
\arguments{
\item{t}{Point where tail probability is evaluated (not this function is not vectorized over t)}

\item{p}{Probabilities of each Bernouilli trial}

\item{s}{Score associated with each trial}

\item{lattice}{Lattice size in minimal lattice}

\item{log}{Return log probability}
}
\description{
Let \eqn{X_1,X_2,\ldots X_N} be independent Bernouilli trials
with \eqn{X_i \sim p_i}.
We evaluate the probability
\deqn{P(s_1X_1+s_2X_2\ldots s_NX_N > t)}.
}
\examples{
poibinSA(300, rep(0.01, 10000))
pbinom(299, 10000, 0.01, lower.tail = F, log.p = T)
}

