% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/potentials.R
\name{potentials}
\alias{potentials}
\title{Potentials}
\usage{
potentials(dfg)
}
\arguments{
\item{dfg}{discrete factor graph object}
}
\value{
A list of current factors
}
\description{
Potentials
}

