% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/mgf.R
\name{saddlePointTails}
\alias{saddlePointTails}
\title{Saddlepoint estimation of tail probabilities}
\usage{
saddlePointTails(t, dfg, facPotFg)
}
\arguments{
\item{t}{value at which to evaluate tail}

\item{dfg}{Probalistic graphical model specifying null model}

\item{facPotFg}{Foreground model}
}
\description{
Saddlepoint estimation of tail probabilities
}

