% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.dfg}
\alias{simulate.dfg}
\title{Simulate from DFG}
\usage{
\method{simulate}{dfg}(x, n)
}
\arguments{
\item{n}{Number of samples}

\item{dfg}{Discrete factor graph object}
}
\description{
Simulate from DFG
}

