% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate
Simulate from a distribution}
\usage{
simulate(x, ...)
}
\arguments{
\item{x}{Distribution from which to simulate.}

\item{...}{Arguments to be passed to methods. Typically \code{n} number of simulations will be provided.}
}
\description{
Simulate
Simulate from a distribution
}

