% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/tail.R
\name{tailIS}
\alias{tailIS}
\title{ImportanceSampling for tail estimation}
\usage{
tailIS(x = NULL, n, q = NULL, alpha = 0.5, dfg, facPotFg)
}
\arguments{
\item{x}{points to evaluate tail probabilities in}

\item{n}{number of samples}

\item{q}{find quantiles instead of probabilities see details}

\item{alpha}{Tuning parameter, alpha=0 is naive sampling the higher alpha the more extreme observations}

\item{dfg}{Probabilistic graphical model specifying null model}

\item{facPotFg}{Foreground model}
}
\value{
A dataframe with columns, x, tail estimate and confidence intervals
}
\description{
ImportanceSampling for tail estimation
}

