% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{EM training}
\usage{
train(data, dfg, threshold = 1e-09, iter.max = 200, dataList = list(),
  verbose = FALSE)
}
\arguments{
\item{data}{dataframe or matrix with observed data. The mapping between columns and rows will be performed automatically. 
NA will be interpreted as a missing variable}

\item{dfg}{discrete factor graph object}

\item{threshold}{Stop training when difference in likelihood between two iterations is below threshold}

\item{iter.max}{Maximal number of iterations of the EM-algorithm.}

\item{dataList}{provides support for partially observed data. See online documentation.}

\item{verbose}{A TRUE/FALSE statement enabling console output of information about the training process}
}
\value{
A discrete factor graph object with updated potentials
}
\description{
EM training
}

