% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.properties.R
\name{trussness}
\alias{graph.trussness}
\alias{trussness}
\title{K-truss decomposition of graphs}
\usage{
trussness(graph)
}
\arguments{
\item{graph}{The input graph ,it is undirected and unweighted}
}
\value{
Numeric vector of integer numbers giving the trussness of each
edge.
}
\description{
The k-truss of graph is a maximal subgraph in which each edge participates in least
k-2 triangles. The trussness of a vertex is k if it belongs to  k-truss but not
to the (k+1)-truss.
}
\details{
The k-truss of graph is a maximal subgraph in which each edge participates in least
k-2 triangles.The trusses of a graph form layers: the (k+1)-truss is always a
subgraph of the k-truss.

This function calculates the trussness for each edge using algorithm 2 in the reference.
}
\examples{

g <- make_ring(10)
g <- add_edges(g, c(1,2, 2,3, 1,3))
trussness(g) 		

}
\author{
Ruchi Goyal \email{ruchigoyal022@gmail.com}

Sana Zehra \email{sanazehra.mcs.du.2014@gmail.com}

Sharanjit kaur \email{skaurjit@yahoo.com}

Vasudha Bhatnagar \email{vasudhabhatnagar@gmail.com}
}
\references{
Wang, Jia, and James Cheng. "Truss decomposition in massive networks." 
Proceedings of the VLDB Endowment 5.9 (2012): 812-823
}
\keyword{graphs}

