% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/potentials.R
\name{update.potential}
\alias{update.potential}
\title{update.potential}
\usage{
\method{update}{potential}(obj, expCounts)
}
\description{
Default row normalization update given expectation counts
}

