% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/potentials.R
\name{betaPotential}
\alias{betaPotential}
\title{Beta Potential
Initialize a beta distributed potential with random starting point}
\usage{
betaPotential(dim = c(1, 100))
}
\arguments{
\item{dim}{A vector with dimensions of potential}
}
\description{
Beta Potential
Initialize a beta distributed potential with random starting point
}

