% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expect.R
\name{expect.dfg}
\alias{expect.dfg}
\title{Expectations in DFG}
\usage{
\method{expect}{dfg}(x, facScores, ...)
}
\arguments{
\item{x}{dfg object}

\item{facScores}{A list of matrices similar to factor potentials indicating a function which expectation should be evaluated}

\item{...}{Arguments to be passed to methods.}
}
\description{
Expectations in DFG
}

