% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facExpectations.R
\name{facExpectations}
\alias{facExpectations}
\title{Calculate expection counts}
\usage{
facExpectations(data, dfg, dataList = list())
}
\arguments{
\item{data}{dataframe or matrix with observed data. The mapping between columns and rows will be performed automatically. NA will be interpreted as a missing variable}

\item{dfg}{discrete factor graph object}

\item{dataList}{provides support for partially observed data. See online documentation.}
}
\value{
A list of matrices containing the expectation counts for each factor
}
\description{
Calculate expection counts
}

