% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{kl}
\alias{kl}
\title{Kullback-Leibler divergence}
\usage{
kl(dfg1, dfg2)
}
\arguments{
\item{dfg1}{Discrete factor graph object describing background distribution}

\item{dfg2}{Discrete factor graph object describing foreground distribution}
}
\description{
Calculates Kullback-leibler divergence KL(dfg1, dfg2) 
between two factor grahs with identical structure. Beware that if for some potential f, and some
variable configuration, x. f(x) > 0 in dfg1 and f(x) = 0 in dfg2, the KL divergence is infinite, 
here however, we ignore such entries and put f(x) = 0 in dfg1.
}

