% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mps.R
\name{mps}
\alias{mps}
\title{Most Probable State}
\usage{
mps(data, dfg, dataList = list())
}
\arguments{
\item{data}{dataframe or matrix with observed data. The mapping between columns and rows will be performed automatically. NA will be interpreted as a missing variable}

\item{dfg}{discrete factor graph object}

\item{dataList}{provides support for partially observed data. See online documentation.}
}
\value{
A dataframe with a column for each variable and the most probable state for each observation
}
\description{
Most Probable State
}

