% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomialPotential.R
\name{multinomialPotential}
\alias{multinomialPotential}
\title{Multinomial Potential}
\usage{
multinomialPotential(dim = c(1, 5), mat = NULL, independent = FALSE,
  pseudocount = 0L, noopt = FALSE)
}
\arguments{
\item{dim}{A vector with dimensions of potential}

\item{mat}{Optionally give fixed matrix potential}

\item{independent}{The two nodes are independent, i.e. rows are identical. This is mainly useful for encoding additional indepedence assumptions in a fixed graph structure.}

\item{pseudocount}{Add pseudocount to each cell. Default is 0.}

\item{noopt}{Hold parameters fixed}
}
\description{
Initialize a multinomial potential with random starting point
}

