% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multinomialPotential.R
\name{multinomialPotential}
\alias{multinomialPotential}
\title{Multinomial Potential}
\usage{
multinomialPotential(dim = c(1, 5))
}
\arguments{
\item{dim}{A vector with dimensions of potential}
}
\description{
Initialize a multinomial potential with random starting point
}

