% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/potentials.R
\name{potentials}
\alias{potentials}
\title{Potentials}
\usage{
potentials(dfg)
}
\arguments{
\item{dfg}{discrete factor graph object}
}
\value{
A list of current factors
}
\description{
Potentials
}
\examples{
varDim <- rep(2,2)
facPot <- list(matrix(c(0.7,0.3),1,2),
               matrix(c(0.75,0.25,0.25,0.75),2,2))
facNbs <- c(list(c(1L)),
            list(c(1L,2L)))
mydfg <- dfg(varDim, facPot, facNbs, varNames = c('x', 'y'))

potentials(mydfg)
}

