% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.dfg}
\alias{simulate.dfg}
\title{Simulate from DFG}
\usage{
\method{simulate}{dfg}(dfg, n)
}
\arguments{
\item{dfg}{Discrete factor graph object}

\item{n}{Number of samples}
}
\description{
Simulate from DFG
}

