% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.dfg}
\alias{simulate.dfg}
\title{Simulate from DFG}
\usage{
\method{simulate}{dfg}(x, nsim, ...)
}
\arguments{
\item{x}{Discrete factor graph object}

\item{nsim}{Number of samples}

\item{...}{Arguments to be passed to methods.}
}
\description{
Simulate from DFG
}

