% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail.R
\name{tailSaddle}
\alias{tailSaddle}
\title{Saddlepoint approximation for tail estimation}
\usage{
tailSaddle(x, dfg1, dfg2, lattice = 0, itermax = 100, tolerance = 0.001)
}
\arguments{
\item{x}{points to evaluate tail probabilities in}

\item{dfg1}{dfg object specifying null model}

\item{dfg2}{dfg object specifying foreground model}

\item{lattice}{correct for scores only taking values on lattice}

\item{itermax}{maximal number of iterations for finding saddlepoint for each x}

\item{tolerance}{tolerance of deviation of mean from x under the tilted distribution}
}
\value{
A dataframe with columns, x, tail estimate and confidence intervals
}
\description{
Saddlepoint approximation for tail estimation
}

