% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{EM training}
\usage{
train(data, dfg, optim = NULL, optimFun = NULL, threshold = 1e-09,
  iter.max = 200, dataList = list(), verbose = FALSE)
}
\arguments{
\item{data}{dataframe or matrix with observed data. The mapping between columns and rows will be performed automatically.
NA will be interpreted as a missing variable}

\item{dfg}{discrete factor graph object}

\item{optim}{character vector with an optimization function for each potential. The built-in functions are
"row" optimizing a multinomial conditional distribution
"norm" optimizing a discretized normal conditional distribution
"linreg" optizing a normal linear regression of x2 on x1}

\item{optimFun}{A named list with additional optimization functions.
Refer to the optimization function by entry name in "optim".}

\item{threshold}{Stop training when difference in likelihood between two iterations is below threshold}

\item{iter.max}{Maximal number of iterations of the EM-algorithm.}

\item{dataList}{provides support for partially observed data. See online documentation.}

\item{verbose}{A TRUE/FALSE statement enabling console output of information about the training process}
}
\value{
A discrete factor graph object with updated potentials
}
\description{
EM training
}

